stupid bandaid fix
in the files here I renamed rotator to ELLrotator because I thought they would conflict with the APT motors.
But they're in separate files, so they don't conflict.
Now, though, the classes have new names, and the code doesn't work with both the elliptec library and this local clone at the same time.
The fix is to go through the local elliptec subfolder and fix everything to be the same as in the official library.